// ccr-ui/src/types/generated/ssh/SshHostConfigDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SshHostConfigDto = { id: string | null, name: string | null, host: string, port: number | null, user: string | null, identity_file: string | null, remote_home: string | null, };